# BackEnd/internals/models/admin.go
package models

// Admin represents the admin user model
type Admin struct {
	ID       int    `json:"id"`
	Username string `json:"username"`
	Password string `json:"-"` // Password is hidden from JSON response
	Email    string `json:"email"`
}

// AdminLogin is used for login credentials
type AdminLogin struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
}

// AdminCreate is used for admin creation
type AdminCreate struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
	Email    string `json:"email" binding:"required,email"`
}

// AdminResponse is used for login response
type AdminResponse struct {
	ID       int    `json:"id"`
	Username string `json:"username"`
	Email    string `json:"email"`
	Token    string `json:"token"`
}